requirements_sora.txt
openai>=1.12.0
requests>=2.31.0
opencv-python>=4.8.0
imageio-ffmpeg>=0.4.9
Pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
colorama>=0.4.6
flask>=3.0.0
flask-cors>=4.0.0
pyyaml>=6.0.1
